Extracted Admin's search behaviour from component to service
ref https://linear.app/tryghost/issue/MOM-1/poc-using-existing-admin-search

- pre-requisite to exposing the search behaviour to the editor for internal linking
