+++ ghost/admin/app/components/gh-search-input.hbs
@@ -1,7 +1,7 @@
 <div ...attributes>
     <div class="ember-power-select-search">
         <PowerSelect
-            @search={{perform this.searchTask}}
+            @search={{perform this.search.searchTask}}
             @onChange={{this.openSelected}}
             @onClose={{this.onClose}}
             @placeholder="Search site"
